doc-requirements.txt
# Requirements for building docs
# Check these dependencies against doc/conf.py
-r dev-requirements.txt
-r test-requirements.txt
sphinx>=1.4
numpydoc
matplotlib
texext
nb2plots
sklearn
sphinx_rtd_theme
